render-layout-lambda-python/requirements.txt
boto3==1.28.38
flask==2.3.3
pillow==10.0.1
requests==2.31.0
